Stripe as source, fixed plugin, added shopping cart
As we now want to do dropshipping and not just link to external sites I needed a payment system. For this I used Stripe. Stripe can hold products and SKUs (variants of a product) which includes details, prices, currency, metadata, ect. As a result I could port all the information that was being stored in contentful to Stripe. However I found that the current plugin for this had a fault with the auth settings for pulling remote images (not hosted on stripe). While this is being reviewed I am including it locally to keep moving forward. I have also added a shopping cart which uses the gatsby-browser file to wrap every page in a permeant React context for shopping cart items. This is more or less working now.