Fade carousel now horizontally centered and even more robust
Previously we capped the width at 400px and the height at 480px. This means that for any image that doesn't have those exact proportions you would get growth in one axis even after the other one had capped out. This caused strange behaviour and made horizontally centering the images quite hard. Now the max width is based on the max height which could be abused for images with an aspect ratio greater than 1 (those that are wider than they are tall) however this is rarely if ever the case. The method for horizontally centering is a bit ugly which is sad, however for now it works. I also discovered that images cropped on the left require too much finessing to bother with for now... You need to have objectFit be cover, not contain for those, and you can only have a single image of this type that is pinned to the left side of the site.