Enhance NavBar to have offscreen mode
When reading a blog post the top navbar is unnecessary as you are unlikely to need those links as frequently as when browsing for clothes. 

To maximise focus on the content the navbar will hide itself as you start scrolling down.  This can be enabled via  the prop `offscreenNavBar` that is part of the Layout component.