Update React Helmet and shallow check SEO component props
React helmet did a deep equality check in should component update. When using the meta tags as children this caused a recursive loop that crashed the app (for whatever reason). The fix was implemented in version 6 which is still unfortunately in beta. I also wrapped the SEO component in react memo to be extra safe.