Shopping cart icon fix and safety net for closing the cart modal
The icon for the cart total was not inline. It is now inline-flex. 

I have also implemented an area that you can click when the shopping cart modal is open that will not trigger other links. The reasoning is that if you click just outside of the modal you most likely meant to close it, and not to press the link that is obscured by the modal. Any strange behaviour with position sticky was documented in the code.