Shopping cart container works at most screen widths and heights
Due to the shopping cart being attached to the navbar which is position sticky, we experienced problems at smaller screen sizes. If the modal went off the screen there was no way to scroll down to it as it was essentially fixed. Also at smaller screen widths text would wrap to the next line causing items to become messy. 

We now have:
* Max height that prevents the container from going over the viewport
* A scroll container that makes all content inside it scrollable while allowing for an escape hatch for items inside it that want to position themselves relative to the parent container.
* The height of the shopping cart psuedo-hardcoded (using css variables to give context about the amount of items and therefore what the height should be). Without this the inner scroll container can't work (it bases its height on its parent).
* A better box shadow (inspired by youtube)
* The triangle on top of the modal has a "border" (actually just two psuedo elements on top of each other with one slightly bigger than the other).

This will most likely be reverted because it feels like a stack of cards (Hardcoding the height is very error prone and hacky). Most likely i will implement another solution that is an item preview page for smaller devices (This is far simpler).