Modify gatsby-source-stripe to pull down SKU images
The current implementation of this plugin assumes that any image hosted on Stripe is private and therefore requires an authorisation header. This is not actually the case and when you provide the auth header when it isn't needed the request fails. So I've modified it to now work for my use case, although there are still some issues in this version as well (It assumes all images are public - they are not).