Shopping cart much simpler - Full screen for small screen height
The previous solution was too brittle as I had to dynamically calculate the height of the content. Now the nav is simply the height of the content at large screen sizes and the height of the screen at small screen sizes. This also makes more sense as it allows me to show the Checkout button all the time (not yet implemented)