Drop shadow of NavBar takes precedence over children
By attaching a drop-shadow directly to the navbar it would be put behind any child content that happened to move over it. In the case of the full screen shopping cart this happened and looked strange. If you place it on a psudo-element than it can have a z-index higher than the child content. Also implemented position/margin based CSS that was previously in the shopping-cart component - this made it less reusable and it shouldn't have been there.