Allow state of useMedia hook to be updated outside custom hook
Due to the app being server render the useMedia hook when initially called can't know the size of the window. You can provide sensible defaults most of the time, but when the default also depends on the screen width you are a bit hosed. So we now expose the set method from the custom hook so that we can update it at any time. This is useful when combined with useEffect, which allows us to measure the screen width as soon as the content is rendered