Netlify function to create payment session
Stripe Checkout client had many problems, including a reduced feature set on the dashboard, which caused me to move away from it. In order to move away from Stripe checkout client to Stripe checkout server I needed to more or less implement the creation of a session myself (this was previously handled for me). This was done using netlify functions.