Item Page created from products, size chart modal added, size dropdown
Previously Item page generation was informed via SKUs. This was incorrect as you would then be generating a page for every size variation of a product. Instead we generate one page per product and have the SKUs that apply to that product inform the content of the page (e.g. what sizes are available). To help with size selection a dropdown has been added that is responsive and accessible (thanks to @reach/ui). The majority of the styling has also been completed, with out of stock being colored differently and featuring a bade. Finally I also added an accessible modal (again thanks to @reach/ui) that will contain the sizing chart.