Wrapper around promise based APIs, session creation in shopping cart
To allow async await to be used easily with fetch we wrap the promises returned by fetch in "to". This will return an array where the first element is the error if it occured and the second is the result. This allows me to not have to wrap each API call in a try catch block. The Shopping cart will use the SKUs selected to generate a session and redirect to checkout once done.