Item metadata pulled down, can select buy without size, styling updates
When you click buy it will now open up the size menu if you haven't already opened it up. If you click on a size it will automatically add it to your cart. The different types of pockets, material and supplier are now pulled down as metadata. Large amount of styling changes to menus.