Add checkout page
The current shopping cart will act as more of a quick view, only allowing you to change quantities if you need to (more for peace of mind by the consumer - you don't want to be forced to click checkout to remove items). Other than that it gives minimal details. The checkout page will allow you to remove items completely, tells you the supplier for each item, and it will soon tell you the shipping time and return policies of each supplier as well as eventually allow you to login.