Fixed no scroll bug
In order to lock the scroll when opening the shopping cart I need to remove the scroll on the body. When the shopping cart closes it allows scrolling again. This was fine until we made the link take you to the checkout page. You would get to the page and not be allowed to scroll because we forgot to remove the style.