Pass on size and metadata to any component ingesting shopping cart
Previously I appended the size to the name of each item in the shopping cart. This was bad because it made the two pieces of info hard to seperate. The metadata is also useful because it contains things like the supplier and how the images are cropped (used on checkout page). Also finally added a key to each element created by the map function.