Increase clickable area of menu button,attempt to fix cart placement bug
I think apple recommends something like 46px by 46px if you want the users to press something with their fingers, the current menu button was much less than that and to make it worse was right near the logo which is a link. 

There is a pesky bug with the nav layout on smaller screens not adjusting for some unknown reason, I'm trying to force Gatsby to hydrate data attributes properly.

Finally passed down close callback to Shopping Cart component