Create reusable PriceDisplay component, add to both index and item page
Quite a few pages may end up needing the discounted price, so I have refactored this functionality into it's own reusable component. This made adding it to the item page much simpler and less error prone. Also refactored the index page to be nicer to look at, plus I can now accept a query string that does or does not have a question mark seperator at the beginning of it.