Added purchase success and cancel pages and emptied cart on success
Would have been bad to get a purchase from a user only to have them redirect to a non-existent page. We now have a plain success and cancel page. On the success page we clear the cart because you don't need it anymore. To avoid a race condition where we hydrate on full page refresh in the context component I now have a check to see if we are on the success page before we hydrate.